font/clientes con botones de agregar, editar y eliminar, se utilizo bootstrap 5 para los iconos y sweetalert2 para las alertas

+++ client/src/app/component/clientes/vista-clientes.component.ts
@@ -1,20 +1,30 @@
 import { Component, OnInit } from '@angular/core';
 import { AuthService } from '../../services/auth.service';
 import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
-import { Router } from '@angular/router';
 import { ClientesService } from '../../services/clientes.service';
+import { FormsModule } from '@angular/forms';
+import Swal from 'sweetalert2';
 
+interface Cliente {
+  id_cliente: number;
+  nombre: string;
+  correo: string;
+  telefono: string;
+  fecha_registro: string;
+}
 @Component({
   selector: 'app-vista-clientes',
-  imports: [CommonModule],
+  standalone: true,
+  imports: [CommonModule, FormsModule],
   templateUrl: './vista-clientes.component.html',
   styleUrl: './vista-clientes.component.scss'
 })
+
+
 export class VistaClientesComponent implements OnInit {
-  clientes: any[] = []
+  clientes: Cliente[] = []
   errorMessage: string = ''
-  nuevoCliente = {nombre: '', correo: '', telefono: ''}
+  nuevoCliente = {id_cliente: 0, nombre: '', correo: '', telefono: ''}
   constructor(private clienteService: ClientesService ){}
 
   ngOnInit(){
@@ -25,6 +35,7 @@ export class VistaClientesComponent implements OnInit {
     this.errorMessage = ''
     this.clienteService.getClientes().subscribe({
       next: (data) => {
+        console.log('Clientes recibidos:', data);
       this.clientes = data;
       },
       error: (error) => {
@@ -39,12 +50,23 @@ export class VistaClientesComponent implements OnInit {
       }
     });
   }
+ agregarCliente() {
+    // Si existe un id, se supone que se trata de una edición.
+    if (this.nuevoCliente.id_cliente) {
+      this.clienteService.updateCliente(this.nuevoCliente.id_cliente, this.nuevoCliente).subscribe(() => {
+        this.obtenerClientes();
+        this.limpiar();
+        Swal.fire('Actualizado', 'El cliente ha sido actualizado.', 'success')
+      });
+    } else {
+      // Si no existe id, se crea un nuevo cliente.
+      this.clienteService.addCliente(this.nuevoCliente).subscribe(() => {
+        this.obtenerClientes();
+        this.limpiar();
+        Swal.fire('Exito', 'El cliente ha sido creado.', 'success');
 
-  agregarCliente() {
-    this.clienteService.addCliente(this.nuevoCliente).subscribe(() => {
-      this.obtenerClientes();
-      this.nuevoCliente = { nombre: '', correo: '', telefono: '' };
-    });
+      });
+    }
   }
 
   editarCliente(cliente: any) {
@@ -54,15 +76,36 @@ export class VistaClientesComponent implements OnInit {
   actualizarCliente(id: number) {
     this.clienteService.updateCliente(id, this.nuevoCliente).subscribe(() => {
       this.obtenerClientes();
-      this.nuevoCliente = { nombre: '', correo: '', telefono: '' };
+      this.nuevoCliente = { id_cliente: 0,nombre: '', correo: '', telefono: '' };
     });
   }
 
   eliminarCliente(id: number) {
-    this.clienteService.deleteCliente(id).subscribe(() => {
-      this.obtenerClientes();
-    });
+    Swal.fire({
+      title: '¿Esta seguro?',
+      text: '¿Esta seguro de eliminar al cliente?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Si',
+      cancelButtonText: 'No'
+    }).then((result)=>{
+      if(result.isConfirmed){
+      // si decidio eliminar el usuario
+      this.clienteService.deleteCliente(id).subscribe(() => {
+        this.obtenerClientes();
+        Swal.fire('Eliminado', 'El cliente ha sido eliminado.', 'success');
+      })
+      }else if (result.dismiss === Swal.DismissReason.cancel) {
+        // si da clic en no se cancela la eliminación.
+        Swal.fire('Cancelado', 'La eliminación ha sido cancelada.', 'error');
+      }
+    })
+ 
   }
+  limpiar() {
+    this.nuevoCliente = { id_cliente: 0, nombre: '', correo: '', telefono: '' };
+  }
+
 
 
 }
